backend-python/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-generativeai
PyPDF2
pdfplumber
python-docx
python-pptx
Pillow
pytesseract
faster-whisper
moviepy
torch
